Issue #3 Add send email
I have added the option to send emails, where the user is asked for the recipient email address, the email subject line and the email's contents. The user will also be able to view their sent emails. 

In this change I also added a check_yes_no function, changed the read_email function to allow for different outputs when dealing with an inbox or an outbox and, changed the parameters in some functions from inbox to folder.